medical_app/requirements.txt
Django==1.10.5
djangorestframework==3.3.0
